Validation result datatypes and tests restructure
