+++ tests/integration/schema/test_link_spec.py
