+++ tests/integration/schema/test_path_params.py
