+++ tests/integration/validation/test_petstore.py
@@ -3,27 +3,20 @@ import pytest
 from datetime import datetime
 from base64 import b64encode
 from uuid import UUID
-from six import iteritems, text_type
+from six import text_type
 
 from openapi_core.extensions.models.models import BaseModel
 from openapi_core.schema.media_types.exceptions import (
     InvalidContentType, InvalidMediaTypeValue,
 )
-from openapi_core.schema.media_types.models import MediaType
-from openapi_core.schema.operations.models import Operation
 from openapi_core.schema.parameters.exceptions import (
     MissingRequiredParameter, InvalidParameterValue, EmptyParameterValue,
 )
-from openapi_core.schema.parameters.models import Parameter
-from openapi_core.schema.paths.models import Path
-from openapi_core.schema.request_bodies.models import RequestBody
-from openapi_core.schema.responses.models import Response
 from openapi_core.schema.schemas.enums import SchemaType
 from openapi_core.schema.schemas.exceptions import InvalidSchemaValue
-from openapi_core.schema.schemas.models import Schema
 from openapi_core.schema.servers.exceptions import InvalidServer
-from openapi_core.schema.servers.models import Server, ServerVariable
 from openapi_core.shortcuts import create_spec
+from openapi_core.validation.request.datatypes import RequestParameters
 from openapi_core.validation.request.validators import RequestValidator
 from openapi_core.validation.response.validators import ResponseValidator
 from openapi_core.wrappers.mock import MockRequest, MockResponse
@@ -59,151 +52,6 @@ class TestPetstore(object):
     def response_validator(self, spec):
         return ResponseValidator(spec)
 
-    def test_spec(self, spec, spec_dict):
-        url = 'http://petstore.swagger.io/v1'
-        assert spec.info.title == spec_dict['info']['title']
-        assert spec.info.version == spec_dict['info']['version']
-
-        assert spec.get_server_url() == url
-
-        for idx, server in enumerate(spec.servers):
-            assert type(server) == Server
-
-            server_spec = spec_dict['servers'][idx]
-            assert server.url == server_spec['url']
-            assert server.default_url == url
-
-            for variable_name, variable in iteritems(server.variables):
-                assert type(variable) == ServerVariable
-                assert variable.name == variable_name
-
-                variable_spec = server_spec['variables'][variable_name]
-                assert variable.default == variable_spec['default']
-                assert variable.enum == variable_spec.get('enum')
-
-        for path_name, path in iteritems(spec.paths):
-            assert type(path) == Path
-            assert path.name == path_name
-
-            for http_method, operation in iteritems(path.operations):
-                operation_spec = spec_dict['paths'][path_name][http_method]
-
-                assert type(operation) == Operation
-                assert operation.path_name == path_name
-                assert operation.http_method == http_method
-                assert operation.operation_id is not None
-                assert operation.tags == operation_spec['tags']
-
-                responses_spec = operation_spec.get('responses')
-
-                for http_status, response in iteritems(operation.responses):
-                    assert type(response) == Response
-                    assert response.http_status == http_status
-
-                    response_spec = responses_spec[http_status]
-
-                    if not response_spec:
-                        continue
-
-                    # @todo: test with defererence
-                    if '$ref' in response_spec:
-                        continue
-
-                    description_spec = response_spec['description']
-
-                    assert response.description == description_spec
-
-                    for mimetype, media_type in iteritems(response.content):
-                        assert type(media_type) == MediaType
-                        assert media_type.mimetype == mimetype
-
-                        content_spec = response_spec['content'][mimetype]
-
-                        example_spec = content_spec.get('example')
-                        assert media_type.example == example_spec
-
-                        schema_spec = content_spec.get('schema')
-                        assert bool(schema_spec) == bool(media_type.schema)
-
-                        if not schema_spec:
-                            continue
-
-                        # @todo: test with defererence
-                        if '$ref' in schema_spec:
-                            continue
-
-                        assert type(media_type.schema) == Schema
-                        assert media_type.schema.type.value ==\
-                            schema_spec['type']
-                        assert media_type.schema.required == schema_spec.get(
-                            'required', [])
-
-                    for parameter_name, parameter in iteritems(
-                            response.headers):
-                        assert type(parameter) == Parameter
-                        assert parameter.name == parameter_name
-
-                        headers_spec = response_spec['headers']
-                        parameter_spec = headers_spec[parameter_name]
-                        schema_spec = parameter_spec.get('schema')
-                        assert bool(schema_spec) == bool(parameter.schema)
-
-                        if not schema_spec:
-                            continue
-
-                        # @todo: test with defererence
-                        if '$ref' in schema_spec:
-                            continue
-
-                        assert type(parameter.schema) == Schema
-                        assert parameter.schema.type.value ==\
-                            schema_spec['type']
-                        assert parameter.schema.format ==\
-                            schema_spec.get('format')
-                        assert parameter.schema.required == schema_spec.get(
-                            'required', [])
-
-                request_body_spec = operation_spec.get('requestBody')
-
-                assert bool(request_body_spec) == bool(operation.request_body)
-
-                if not request_body_spec:
-                    continue
-
-                assert type(operation.request_body) == RequestBody
-                assert bool(operation.request_body.required) ==\
-                    request_body_spec.get('required', False)
-
-                for mimetype, media_type in iteritems(
-                        operation.request_body.content):
-                    assert type(media_type) == MediaType
-                    assert media_type.mimetype == mimetype
-
-                    content_spec = request_body_spec['content'][mimetype]
-                    schema_spec = content_spec.get('schema')
-                    assert bool(schema_spec) == bool(media_type.schema)
-
-                    if not schema_spec:
-                        continue
-
-                    # @todo: test with defererence
-                    if '$ref' in schema_spec:
-                        continue
-
-                    assert type(media_type.schema) == Schema
-                    assert media_type.schema.type.value ==\
-                        schema_spec['type']
-                    assert media_type.schema.format ==\
-                        schema_spec.get('format')
-                    assert media_type.schema.required == schema_spec.get(
-                        'required', False)
-
-        if not spec.components:
-            return
-
-        for schema_name, schema in iteritems(spec.components.schemas):
-            assert type(schema) == Schema
-
     def test_get_pets(self, spec, response_validator):
         host_url = 'http://petstore.swagger.io/v1'
         path_pattern = '/v1/pets'
@@ -219,13 +67,13 @@ class TestPetstore(object):
         parameters = request.get_parameters(spec)
         body = request.get_body(spec)
 
-        assert parameters == {
-            'query': {
+        assert parameters == RequestParameters(
+            query={
                 'limit': 20,
                 'page': 1,
                 'search': '',
             }
-        }
+        )
         assert body is None
 
         data_json = {
@@ -255,13 +103,13 @@ class TestPetstore(object):
         parameters = request.get_parameters(spec)
         body = request.get_body(spec)
 
-        assert parameters == {
-            'query': {
+        assert parameters == RequestParameters(
+            query={
                 'limit': 20,
                 'page': 1,
                 'search': '',
             }
-        }
+        )
         assert body is None
 
         data_json = {
@@ -301,13 +149,13 @@ class TestPetstore(object):
         parameters = request.get_parameters(spec)
         body = request.get_body(spec)
 
-        assert parameters == {
-            'query': {
+        assert parameters == RequestParameters(
+            query={
                 'limit': 20,
                 'page': 1,
                 'search': '',
             }
-        }
+        )
         assert body is None
 
         data_json = {
@@ -352,14 +200,14 @@ class TestPetstore(object):
         parameters = request.get_parameters(spec)
         body = request.get_body(spec)
 
-        assert parameters == {
-            'query': {
+        assert parameters == RequestParameters(
+            query={
                 'limit': 20,
                 'page': 1,
                 'search': '',
                 'ids': [12, 13],
             }
-        }
+        )
         assert body is None
 
         data_json = {
@@ -390,14 +238,14 @@ class TestPetstore(object):
         parameters = request.get_parameters(spec)
         body = request.get_body(spec)
 
-        assert parameters == {
-            'query': {
+        assert parameters == RequestParameters(
+            query={
                 'limit': 20,
                 'page': 1,
                 'search': '',
                 'tags': ['cats', 'dogs'],
             }
-        }
+        )
         assert body is None
 
         data_json = {
@@ -498,13 +346,13 @@ class TestPetstore(object):
 
         parameters = request.get_parameters(spec)
 
-        assert parameters == {
-            'query': {
+        assert parameters == RequestParameters(
+            query={
                 'limit': None,
                 'page': 1,
                 'search': '',
             }
-        }
+        )
 
         body = request.get_body(spec)
 
@@ -547,14 +395,14 @@ class TestPetstore(object):
 
         parameters = request.get_parameters(spec)
 
-        assert parameters == {
-            'header': {
+        assert parameters == RequestParameters(
+            header={
                 'api_key': self.api_key,
             },
-            'cookie': {
+            cookie={
                 'user': 123,
             },
-        }
+        )
 
         body = request.get_body(spec)
 
@@ -607,14 +455,14 @@ class TestPetstore(object):
 
         parameters = request.get_parameters(spec)
 
-        assert parameters == {
-            'header': {
+        assert parameters == RequestParameters(
+            header={
                 'api_key': self.api_key,
             },
-            'cookie': {
+            cookie={
                 'user': 123,
             },
-        }
+        )
 
         body = request.get_body(spec)
 
@@ -667,14 +515,14 @@ class TestPetstore(object):
 
         parameters = request.get_parameters(spec)
 
-        assert parameters == {
-            'header': {
+        assert parameters == RequestParameters(
+            header={
                 'api_key': self.api_key,
             },
-            'cookie': {
+            cookie={
                 'user': 123,
             },
-        }
+        )
 
         body = request.get_body(spec)
 
@@ -715,14 +563,14 @@ class TestPetstore(object):
 
         parameters = request.get_parameters(spec)
 
-        assert parameters == {
-            'header': {
+        assert parameters == RequestParameters(
+            header={
                 'api_key': self.api_key,
             },
-            'cookie': {
+            cookie={
                 'user': 123,
             },
-        }
+        )
 
         with pytest.raises(InvalidMediaTypeValue):
             request.get_body(spec)
@@ -754,14 +602,14 @@ class TestPetstore(object):
 
         parameters = request.get_parameters(spec)
 
-        assert parameters == {
-            'header': {
+        assert parameters == RequestParameters(
+            header={
                 'api_key': self.api_key,
             },
-            'cookie': {
+            cookie={
                 'user': 123,
             },
-        }
+        )
 
         body = request.get_body(spec)
 
@@ -795,14 +643,14 @@ class TestPetstore(object):
 
         parameters = request.get_parameters(spec)
 
-        assert parameters == {
-            'header': {
+        assert parameters == RequestParameters(
+            header={
                 'api_key': self.api_key,
             },
-            'cookie': {
+            cookie={
                 'user': 123,
             },
-        }
+        )
 
         with pytest.raises(InvalidContentType):
             request.get_body(spec)
@@ -915,11 +763,11 @@ class TestPetstore(object):
 
         parameters = request.get_parameters(spec)
 
-        assert parameters == {
-            'path': {
+        assert parameters == RequestParameters(
+            path={
                 'petId': 1,
             }
-        }
+        )
 
         body = request.get_body(spec)
 
@@ -960,11 +808,11 @@ class TestPetstore(object):
 
         parameters = request.get_parameters(spec)
 
-        assert parameters == {
-            'path': {
+        assert parameters == RequestParameters(
+            path={
                 'petId': 1,
             }
-        }
+        )
 
         body = request.get_body(spec)
 
@@ -1002,11 +850,11 @@ class TestPetstore(object):
 
         parameters = request.get_parameters(spec)
 
-        assert parameters == {
-            'path': {
+        assert parameters == RequestParameters(
+            path={
                 'petId': 1,
             }
-        }
+        )
 
         body = request.get_body(spec)
 
@@ -1032,7 +880,7 @@ class TestPetstore(object):
         parameters = request.get_parameters(spec)
         body = request.get_body(spec)
 
-        assert parameters == {}
+        assert parameters == RequestParameters()
         assert body is None
 
         data_json = ['cats', 'birds']
@@ -1062,7 +910,7 @@ class TestPetstore(object):
 
         parameters = request.get_parameters(spec)
 
-        assert parameters == {}
+        assert parameters == RequestParameters()
 
         with pytest.raises(InvalidMediaTypeValue):
             request.get_body(spec)
@@ -1080,7 +928,7 @@ class TestPetstore(object):
 
         parameters = request.get_parameters(spec)
 
-        assert parameters == {}
+        assert parameters == RequestParameters()
 
         with pytest.raises(InvalidMediaTypeValue):
             request.get_body(spec)
@@ -1098,7 +946,7 @@ class TestPetstore(object):
 
         parameters = request.get_parameters(spec)
 
-        assert parameters == {}
+        assert parameters == RequestParameters()
 
         with pytest.raises(InvalidMediaTypeValue):
             request.get_body(spec)
@@ -1121,7 +969,7 @@ class TestPetstore(object):
         parameters = request.get_parameters(spec)
         body = request.get_body(spec)
 
-        assert parameters == {}
+        assert parameters == RequestParameters()
         assert isinstance(body, BaseModel)
         assert body.name == pet_name
 
@@ -1167,7 +1015,7 @@ class TestPetstore(object):
         parameters = request.get_parameters(spec)
         body = request.get_body(spec)
 
-        assert parameters == {}
+        assert parameters == RequestParameters()
         assert isinstance(body, BaseModel)
         assert body.created == created
         assert body.name == pet_name
@@ -1214,7 +1062,7 @@ class TestPetstore(object):
         parameters = request.get_parameters(spec)
         body = request.get_body(spec)
 
-        assert parameters == {}
+        assert parameters == RequestParameters()
         assert isinstance(body, BaseModel)
         assert body.created == datetime(2016, 4, 16, 16, 6, 5)
         assert body.name == pet_name
@@ -1262,7 +1110,7 @@ class TestPetstore(object):
         with pytest.raises(InvalidMediaTypeValue):
             request.get_body(spec)
 
-        assert parameters == {}
+        assert parameters == RequestParameters()
 
         code = 400
         message = 'Bad request'
